requirements.txt
# Core dependencies
python-dateutil>=2.8.2
pyyaml>=6.0

# Document processing
# tesseract-ocr  # System dependency
# apache-tika  # Optional - requires Java
# pdf2image>=1.16.0

# NLP and ML
spacy>=3.5.0
transformers>=4.25.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3  # or faiss-gpu for GPU support

# Vector databases
# qdrant-client>=1.1.1
# elasticsearch>=8.6.0

# RAG and LLM
langchain>=0.0.207
guardrails-ai>=0.1.0

# Security
cryptography>=38.0.0

# Testing
pytest>=7.0.0
pylint>=2.17.0
